backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pymupdf==1.24.4
spacy==3.7.4
sentence-transformers==2.7.0
faiss-cpu==1.8.0
networkx==3.3
groq==0.9.0
scikit-learn==1.4.2
numpy==1.26.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.3